refactor: move update-branch types and functions from apply to update-pr-branch
These types and functions were only consumed by update-pr-branch after
the extraction in the previous commit, so colocate them with their
sole consumer.
